third_party/blink/web_tests/platform/generic/html/tabular_data/table_tbodies-expected.txt

Test behavior of the HTMLTableElement tBodies attribute in cases where there is unusual nesting.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS checkTBodyNesting("col") is 0
PASS checkTBodyNesting("colgroup") is 0
PASS checkTBodyNesting("div") is 0
PASS checkTBodyNesting("form") is 0
PASS checkTBodyNesting("script") is 0
PASS checkTBodyNesting("table") is 0
PASS checkTBodyNesting("td") is 0
PASS checkTBodyNesting("tfoot") is 0
PASS checkTBodyNesting("th") is 0
PASS checkTBodyNesting("thead") is 0
PASS checkTBodyNesting("tr") is 0


PASS checkTBodyNesting("tbody") is 1


PASS successfullyParsed is true

TEST COMPLETE
